# modules/apps/app-builder/app-builder-rest-impl/rest-openapi.yaml
components:
    schemas:
        App:
            properties:
                active:
                    type: boolean
                appDeployments:
                    items:
                        $ref: "#/components/schemas/AppDeployment"
                    type: array
                dataDefinitionId:
                    format: int64
                    type: integer
                dataDefinitionName:
                    readOnly: true
                    type: string
                dataLayoutId:
                    format: int64
                    type: integer
                dataListViewId:
                    format: int64
                    type: integer
                dataRecordCollectionId:
                    format: int64
                    type: integer
                dateCreated:
                    format: date-time
                    type: string
                dateModified:
                    format: date-time
                    type: string
                id:
                    format: int64
                    type: integer
                name:
                    additionalProperties:
                        type: object
                    type: object
                scope:
                    type: string
                siteId:
                    format: int64
                    type: integer
                userId:
                    format: int64
                    type: integer
                version:
                    type: string
            type: object
        AppDeployment:
            properties:
                settings:
                    additionalProperties:
                        type: object
                    type: object
                type:
                    type: string
            type: object
info:
    description:
        "A Java client JAR is available for use with the group ID 'com.liferay', artifact ID
        'com.liferay.app.builder.rest.client', and version '1.0.11'."
    license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
    title: "App Builder"
    version: v1.0
openapi: 3.0.1
paths:
    "/apps":
        get:
            operationId: getAppsPage
            parameters:
                - in: query
                  name: active
                  schema:
                      type: boolean
                - in: query
                  name: deploymentTypes
                  schema:
                      items:
                          enum: [productMenu, standalone, widget]
                          type: string
                      type: array
                - in: query
                  name: keywords
                  schema:
                      type: string
                - in: query
                  name: scope
                  schema:
                      type: string
                - in: query
                  name: userIds
                  schema:
                      items:
                          format: int64
                          type: integer
                      type: array
                - in: query
                  name: page
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  schema:
                      type: integer
                - in: query
                  name: sort
                  required: false
                  schema:
                      type: string
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/App"
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: "#/components/schemas/App"
                                type: array
                    description:
                        ""
            tags: ["App"]
    "/apps/{appId}":
        delete:
            operationId: deleteApp
            parameters:
                - in: path
                  name: appId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                204:
                    content:
                        application/json: {}
                        application/xml: {}
                    description:
                        ""
            tags: ["App"]
        get:
            operationId: getApp
            parameters:
                - in: path
                  name: appId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/App"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/App"
                    description:
                        ""
            tags: ["App"]
        put:
            operationId: putApp
            parameters:
                - in: path
                  name: appId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/App"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/App"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/App"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/App"
                    description:
                        ""
            tags: ["App"]
    "/apps/{appId}/deploy":
        put:
            operationId: putAppDeploy
            parameters:
                - in: path
                  name: appId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                200:
                    description:
                        "OK"
            tags: ["App"]
    "/apps/{appId}/undeploy":
        put:
            operationId: putAppUndeploy
            parameters:
                - in: path
                  name: appId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                200:
                    description:
                        "OK"
            tags: ["App"]
    "/data-definitions/{dataDefinitionId}/apps":
        get:
            operationId: getDataDefinitionAppsPage
            parameters:
                - in: path
                  name: dataDefinitionId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: keywords
                  schema:
                      type: string
                - in: query
                  name: scope
                  schema:
                      type: string
                - in: query
                  name: page
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  schema:
                      type: integer
                - in: query
                  name: sort
                  required: false
                  schema:
                      type: string
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/App"
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: "#/components/schemas/App"
                                type: array
                    description:
                        ""
            tags: ["App"]
        post:
            operationId: postDataDefinitionApp
            parameters:
                - in: path
                  name: dataDefinitionId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/App"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/App"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/App"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/App"
                    description:
                        ""
            tags: ["App"]
    "/sites/{siteId}/apps":
        get:
            operationId: getSiteAppsPage
            parameters:
                - in: path
                  name: siteId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: keywords
                  schema:
                      type: string
                - in: query
                  name: scope
                  schema:
                      type: string
                - in: query
                  name: page
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  schema:
                      type: integer
                - in: query
                  name: sort
                  required: false
                  schema:
                      type: string
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/App"
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: "#/components/schemas/App"
                                type: array
                    description:
                        ""
            tags: ["App"]